// frontend/src/types/common/queryBuilder.ts
export enum DataSource {
	METRICS = 'metrics',
	TRACES = 'traces',
	LOGS = 'logs',
}